app/requirements.txt
Django==2.2.6
gunicorn==19.9.0
psycopg2-binary==2.8.3
numpy
pillow==6.2.2
scikit-image
fastai==1.0.51
wandb
tensorboardX==1.6
ffmpeg
ffmpeg-python==0.1.17
opencv-python>=3.3.0.10
youtube_dl